refactor: split bayesian/mcmc.py (691L) into peer modules (#225)
- mcmc.py (270L) — MCMCResult, MCMCForecaster, mcmc_forecast() + re-imports
- mcmc_samplers.py (135L) — log-posteriors + MH sampler
- mcmc_forecast.py (80L) — posterior predictive forecast functions
- mcmc_backends.py (96L) — NumPyro/PyMC backends
- All files under 500 lines (max 270L vs original 691L)
- All existing import paths preserved via re-imports in mcmc.py
- 6 new structural tests, 41/41 total tests pass
